issues #822 and #763 in laravel5 branch
I transfered the requests from frontend to backend used to share notes
and notebooks from GET to POST. This should resolve issue #822.

I also changed the "share notebook" feature in the frontend as mentioned
in issue #763 :
 - a share button is added in the Manage Notebooks modal view which
needed a little refactoring of a few methods (e.g. moving GetUsers
 function from NotebookSidebarController to NotebookService)
 - the propagation of the sharing status of a Notebook to all the notes
included in it is improved : a single request is sent to the backend
 instead of one per note in the previous version.